Added functions to rigid2d.cpp

+++ turtlelib/rigid2d.hpp
@@ -157,7 +157,6 @@ namespace turtlelib
         /// \return a vector in the new coordinate system
         Vector2D operator()(Vector2D v) const;
 
-
         /// \brief invert the transformation
         /// \return the inverse transformation. 
         Transform2D inv() const;
@@ -180,7 +179,7 @@ namespace turtlelib
         /// for a description
         friend std::ostream & operator<<(std::ostream & os, const Transform2D & tf);
     private:
-        int T[3][3];
+        float T[3][3];
     };
